+++ src/games/mod.rs
@@ -1,0 +1,2 @@
+mod tic_tac_toe;
+pub use tic_tac_toe::TicTacToe;
