+++ src/main.rs
@@ -1,4 +1,4 @@
-use minimax_alpha_beta::drivers::*;
+use minimax_alpha_beta::*;
 
 fn main() {
     play_tic_tac_toe_against_computer_with_depth(4, 7);
