+++ src/strategy/mod.rs
@@ -1,0 +1,2 @@
+pub mod alpha_beta_minimax;
+pub mod game_strategy;
